move ImportProxy to lib and write some docs
